scripts/video-processor/requirements.txt
absl-py==2.2.0
accelerate==1.3.0
addict==2.4.0
aiobotocore==2.21.1
aiodns==3.2.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
antlr4-tools==0.2.1
anyio==4.8.0
art==6.4
attrs==25.1.0
axolotl==0.7.0
axolotl-contribs-lgpl==0.0.3
backoff==2.2.1
bcrypt==4.3.0
boto3==1.37.9
botocore==1.37.1
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
coverage==7.6.12
cryptography==44.0.2
cycler==0.12.1
DataProperty==1.1.0
datasets==3.2.0
decorator==5.2.1
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.1
email_validator==2.2.0
evaluate==0.4.1
Faker==37.1.0
fastapi==0.115.11
fastapi-cli==0.0.7
fastcore==1.8.0
ffmpy==0.5.0
filelock==3.17.0
fire==0.7.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.9.0
gcsfs==2024.9.0.post1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
gotrue==2.12.0
gradio==3.50.2
gradio_client==0.6.1
grpcio==1.71.0
grpclib==0.4.7
gym==0.26.2
gym-notices==0.0.8
h11==0.16.0
h2==4.2.0
hf_transfer==0.1.9
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
immutabledict==4.2.0
importlib_resources==6.5.2
iniconfig==2.0.0
inquirerpy==0.3.4
install-jdk==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwt==1.3.1
kiwisolver==1.4.8
langdetect==1.0.9
llvmlite==0.44.0
lm_eval==0.4.7
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mbstrdecoder==1.1.4
mdurl==0.1.2
modal==0.70.5
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.31.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numexpr==2.10.2
numpy==1.26.4
nvidia-ml-py==12.560.30
oauthlib==3.2.2
openai==1.61.1
optimum==1.16.2
orjson==3.10.15
packaging==23.2
pandas==2.2.3
paramiko==3.5.1
pathvalidate==3.2.3
peft==0.14.0
pfzy==0.3.4
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
portalocker==3.1.1
postgrest==1.0.2
prettytable==3.15.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycares==4.5.0
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.1
pytablewriter==1.2.1
pytest==8.3.4
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
realtime==2.4.3
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
responses==0.18.0
rich==13.9.4
rich-toolkit==0.13.2
rouge_score==0.1.2
rpds-py==0.23.1
rsa==4.9
runpod==1.7.7
s3fs==2024.9.0
s3transfer==0.11.4
sacrebleu==2.5.1
safetensors==0.5.2
schedulefree==1.3
scikit-learn==1.4.2
scipy==1.15.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.23.1
setproctitle==1.3.5
setuptools==75.8.0
shellingham==1.5.4
shtab==1.7.1
sigtools==4.0.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlitedict==2.1.0
starlette==0.46.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.2
supafunc==0.9.4
sympy==1.13.1
synchronicity==0.9.11
ta==0.11.0
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
threadpoolctl==3.6.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0
torchao==0.7.0
tqdm==4.67.1
tqdm-loggable==0.2
tqdm-multiprocess==0.0.11
transformers==4.48.3
trl==0.15.0
typeguard==4.4.2
typepy==1.3.4
typer==0.15.1
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing_extensions==4.14.0
tyro==0.9.16
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
wandb==0.19.8
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.1.3
word2number==1.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zstandard==0.22.0
